Refactoring part 1. Fix react deprecation notice. Fix styles. Delete what is not used. Change runts. install  @ton-community/func-js .
