+++ package.json
@@ -45,6 +45,7 @@
   "dependencies": {
     "-": "^0.0.1",
     "@monaco-editor/react": "^4.4.6",
+    "@ton-community/func-js": "^0.5.0",
     "@typescript-eslint/eslint-plugin": "^5.0.0",
     "@typescript-eslint/parser": "^5.0.0",
     "bn.js": "^5.2.1",
