+++ src/index.html
@@ -3,46 +3,9 @@
 <head>
     <meta charset="utf-8"/>
     <title>FunC editor</title>
-    <style>
-        html, body {
-            background-color: #1e1e1e;
-            margin: 0;
-            padding: 0;
-            width: 100vw;
-            height: 100vh;
-            overflow: hidden;
-        }
-
-        body {
-            position: relative;
-        }
-
-        h3 {
-            font-family: 'Roboto Mono';
-        }
-
-        h4 {
-            font-family: 'Roboto Mono';
-        }
-
-        .monaco-tree:focus {
-            outline: none;
-        }
-
-        pre {
-            white-space: pre-wrap;
-            word-wrap: break-word;
-        }
-    </style>
     <link href='https://fonts.googleapis.com/css?family=Roboto Mono' rel='stylesheet'>
 </head>
 <body>
-<h3 style="padding-left: 15px; color: #fff">FunC Editor <small style="color: #aaa">press [shift+enter] to run the code</small></h3>
-<div style="display: inline-block; width: 15%; height: 100vh; position: absolute">
-    <div style="width: 100%; height: calc(100vh - 4em);;" id="tree"></div>
-</div>
-<div style="display: inline-block; width: 85%; height: 100vh; left: 15%; position: absolute">
     <div id="root"></div>
-</div>
 </body>
 </html>
