+++ src/styles/Editor/Preview.ts
@@ -1,0 +1,12 @@
+import styled from "styled-components";
+
+export const Preview = styled.div`
+  background: #181818;
+  width: 50vw;
+  height: 100vh;
+  word-break: break-word;
+  padding: 5px;
+  box-sizing: border-box;
+  color: whitesmoke;
+  overflow-y: auto;
+`
