+++ src/styles/Editor/Wrap.ts
@@ -1,0 +1,5 @@
+import styled from "styled-components";
+
+export const Wrap = styled.div`
+  display: flex;
+`
