+++ src/styles/Editor/index.ts
@@ -1,0 +1,4 @@
+import {Preview} from "./Preview";
+import {Wrap} from "./Wrap";
+
+export {Preview, Wrap};
