+++ src/styles/index.css
@@ -1,0 +1,29 @@
+html, body {
+    background-color: #1e1e1e;
+    margin: 0;
+    padding: 0;
+    width: 100vw;
+    height: 100vh;
+    overflow: hidden;
+}
+
+body {
+    position: relative;
+}
+
+h3 {
+    font-family: 'Roboto Mono', serif;
+}
+
+h4 {
+    font-family: 'Roboto Mono', serif;
+}
+
+.monaco-tree:focus {
+    outline: none;
+}
+
+pre {
+    white-space: pre-wrap;
+    word-wrap: break-word;
+}
